server: isolate linux only functions
