+++ server/container_create_linux.go
@@ -3,20 +3,31 @@
 package server
 
 import (
+	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
 	"runtime"
+	"sort"
 	"strings"
+	"time"
 
 	"github.com/containers/storage/pkg/idtools"
+	dockermounts "github.com/docker/docker/pkg/mount"
+	"github.com/docker/docker/pkg/symlink"
 	"github.com/kubernetes-incubator/cri-o/lib/sandbox"
 	"github.com/kubernetes-incubator/cri-o/oci"
+	"github.com/kubernetes-incubator/cri-o/pkg/annotations"
+	"github.com/kubernetes-incubator/cri-o/pkg/apparmor"
+	"github.com/kubernetes-incubator/cri-o/pkg/storage"
 	"github.com/opencontainers/runc/libcontainer/cgroups"
 	"github.com/opencontainers/runc/libcontainer/devices"
 	rspec "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/opencontainers/runtime-tools/generate"
+	"github.com/pkg/errors"
+	"github.com/projectatomic/libpod/pkg/secrets"
 	"github.com/sirupsen/logrus"
+	"golang.org/x/net/context"
 	"golang.org/x/sys/unix"
 	pb "k8s.io/kubernetes/pkg/kubelet/apis/cri/runtime/v1alpha2"
 )
@@ -235,4 +246,762 @@ func (s *Server) createContainerPlatform(container *oci.Container, infraContaine
 
 	err := <-errc
 	return err
+}
+
+func (s *Server) createSandboxContainer(ctx context.Context, containerID string, containerName string, sb *sandbox.Sandbox, sandboxConfig *pb.PodSandboxConfig, containerConfig *pb.ContainerConfig) (*oci.Container, error) {
+	if sb == nil {
+		return nil, errors.New("createSandboxContainer needs a sandbox")
+	}
+
+	// TODO: simplify this function (cyclomatic complexity here is high)
+	// TODO: factor generating/updating the spec into something other projects can vendor
+
+	// creates a spec Generator with the default spec.
+	specgen, err := generate.New("linux")
+	if err != nil {
+		return nil, err
+	}
+	specgen.HostSpecific = true
+	specgen.ClearProcessRlimits()
+
+	readOnlyRootfs := s.config.ReadOnly
+
+	var privileged bool
+	if containerConfig.GetLinux().GetSecurityContext() != nil {
+		if containerConfig.GetLinux().GetSecurityContext().GetPrivileged() {
+			privileged = true
+		}
+		if privileged {
+			if !sandboxConfig.GetLinux().GetSecurityContext().GetPrivileged() {
+				return nil, errors.New("no privileged container allowed in sandbox")
+			}
+		}
+		if containerConfig.GetLinux().GetSecurityContext().GetReadonlyRootfs() {
+			readOnlyRootfs = true
+		}
+	}
+	specgen.SetRootReadonly(readOnlyRootfs)
+
+	if s.config.ReadOnly {
+		// tmpcopyup is a runc extension and is not part of the OCI spec.
+		// WORK ON: Use "overlay" mounts as an alternative to tmpfs with tmpcopyup
+		// Look at https://github.com/kubernetes-incubator/cri-o/pull/1434#discussion_r177200245 for more info on this
+		options := []string{"rw", "noexec", "nosuid", "nodev", "tmpcopyup"}
+		if !isInCRIMounts("/run", containerConfig.GetMounts()) {
+			mnt := rspec.Mount{
+				Destination: "/run",
+				Type:        "tmpfs",
+				Source:      "tmpfs",
+				Options:     append(options, "mode=0755"),
+			}
+			// Add tmpfs mount on /run
+			specgen.AddMount(mnt)
+		}
+		if !isInCRIMounts("/tmp", containerConfig.GetMounts()) {
+			mnt := rspec.Mount{
+				Destination: "/tmp",
+				Type:        "tmpfs",
+				Source:      "tmpfs",
+				Options:     append(options, "mode=1777"),
+			}
+			// Add tmpfs mount on /tmp
+			specgen.AddMount(mnt)
+		}
+		if !isInCRIMounts("/var/tmp", containerConfig.GetMounts()) {
+			mnt := rspec.Mount{
+				Destination: "/var/tmp",
+				Type:        "tmpfs",
+				Source:      "tmpfs",
+				Options:     append(options, "mode=1777"),
+			}
+			// Add tmpfs mount on /var/tmp
+			specgen.AddMount(mnt)
+		}
+	}
+	mountLabel := sb.MountLabel()
+	processLabel := sb.ProcessLabel()
+	selinuxConfig := containerConfig.GetLinux().GetSecurityContext().GetSelinuxOptions()
+	if selinuxConfig != nil {
+		var err error
+		processLabel, mountLabel, err = getSELinuxLabels(selinuxConfig, privileged)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	containerVolumes, ociMounts, err := addOCIBindMounts(mountLabel, containerConfig, &specgen, s.config.RuntimeConfig.BindMountPrefix)
+	if err != nil {
+		return nil, err
+	}
+
+	volumesJSON, err := json.Marshal(containerVolumes)
+	if err != nil {
+		return nil, err
+	}
+	specgen.AddAnnotation(annotations.Volumes, string(volumesJSON))
+
+	mnt := rspec.Mount{
+		Destination: "/sys/fs/cgroup",
+		Type:        "cgroup",
+		Source:      "cgroup",
+		Options:     []string{"nosuid", "noexec", "nodev", "relatime", "ro"},
+	}
+	// Add cgroup mount so container process can introspect its own limits
+	specgen.AddMount(mnt)
+
+	if err := addDevices(sb, containerConfig, &specgen); err != nil {
+		return nil, err
+	}
+
+	labels := containerConfig.GetLabels()
+
+	if err := validateLabels(labels); err != nil {
+		return nil, err
+	}
+
+	metadata := containerConfig.GetMetadata()
+
+	kubeAnnotations := containerConfig.GetAnnotations()
+	for k, v := range kubeAnnotations {
+		specgen.AddAnnotation(k, v)
+	}
+	for k, v := range labels {
+		specgen.AddAnnotation(k, v)
+	}
+
+	// set this container's apparmor profile if it is set by sandbox
+	if s.appArmorEnabled && !privileged {
+
+		appArmorProfileName := s.getAppArmorProfileName(containerConfig.GetLinux().GetSecurityContext().GetApparmorProfile())
+		if appArmorProfileName != "" {
+			// reload default apparmor profile if it is unloaded.
+			if s.appArmorProfile == apparmor.DefaultApparmorProfile {
+				if err := apparmor.EnsureDefaultApparmorProfile(); err != nil {
+					return nil, err
+				}
+			}
+
+			specgen.SetProcessApparmorProfile(appArmorProfileName)
+		}
+
+	}
+
+	logPath := containerConfig.GetLogPath()
+	sboxLogDir := sandboxConfig.GetLogDirectory()
+	if sboxLogDir == "" {
+		sboxLogDir = sb.LogDir()
+	}
+	if logPath == "" {
+		logPath = filepath.Join(sboxLogDir, containerID+".log")
+	}
+	if !filepath.IsAbs(logPath) {
+		// XXX: It's not really clear what this should be versus the sbox logDirectory.
+		logrus.Warnf("requested logPath for ctr id %s is a relative path: %s", containerID, logPath)
+		logPath = filepath.Join(sboxLogDir, logPath)
+		logrus.Warnf("logPath from relative path is now absolute: %s", logPath)
+	}
+
+	// Handle https://issues.k8s.io/44043
+	if err := ensureSaneLogPath(logPath); err != nil {
+		return nil, err
+	}
+
+	logrus.WithFields(logrus.Fields{
+		"sbox.logdir": sboxLogDir,
+		"ctr.logfile": containerConfig.GetLogPath(),
+		"log_path":    logPath,
+	}).Debugf("setting container's log_path")
+
+	specgen.SetProcessTerminal(containerConfig.Tty)
+	if containerConfig.Tty {
+		specgen.AddProcessEnv("TERM", "xterm")
+	}
+
+	linux := containerConfig.GetLinux()
+	if linux != nil {
+		resources := linux.GetResources()
+		if resources != nil {
+			specgen.SetLinuxResourcesCPUPeriod(uint64(resources.GetCpuPeriod()))
+			specgen.SetLinuxResourcesCPUQuota(resources.GetCpuQuota())
+			specgen.SetLinuxResourcesCPUShares(uint64(resources.GetCpuShares()))
+			specgen.SetLinuxResourcesMemoryLimit(resources.GetMemoryLimitInBytes())
+			specgen.SetProcessOOMScoreAdj(int(resources.GetOomScoreAdj()))
+			specgen.SetLinuxResourcesCPUCpus(resources.GetCpusetCpus())
+			specgen.SetLinuxResourcesCPUMems(resources.GetCpusetMems())
+		}
+
+		var cgPath string
+		parent := defaultCgroupfsParent
+		useSystemd := s.config.CgroupManager == oci.SystemdCgroupsManager
+		if useSystemd {
+			parent = defaultSystemdParent
+		}
+		if sb.CgroupParent() != "" {
+			parent = sb.CgroupParent()
+		}
+		if useSystemd {
+			cgPath = parent + ":" + scopePrefix + ":" + containerID
+		} else {
+			cgPath = filepath.Join(parent, scopePrefix+"-"+containerID)
+		}
+		specgen.SetLinuxCgroupsPath(cgPath)
+
+		if privileged {
+			specgen.SetupPrivileged(true)
+			setOCIBindMountsPrivileged(&specgen)
+		} else {
+			capabilities := linux.GetSecurityContext().GetCapabilities()
+			// Ensure we don't get a nil pointer error if the config
+			// doesn't set any capabilities
+			if capabilities == nil {
+				capabilities = &pb.Capability{}
+			}
+			// Clear default capabilities from spec
+			specgen.ClearProcessCapabilities()
+			for _, defaultCap := range s.config.DefaultCapabilities {
+				capabilities.AddCapabilities = append(capabilities.AddCapabilities, defaultCap)
+			}
+			err = setupCapabilities(&specgen, capabilities)
+			if err != nil {
+				return nil, err
+			}
+		}
+		specgen.SetProcessSelinuxLabel(processLabel)
+		specgen.SetLinuxMountLabel(mountLabel)
+		specgen.SetProcessNoNewPrivileges(linux.GetSecurityContext().GetNoNewPrivs())
+
+		if containerConfig.GetLinux().GetSecurityContext() != nil &&
+			!containerConfig.GetLinux().GetSecurityContext().Privileged {
+			for _, mp := range []string{
+				"/proc/acpi",
+				"/proc/kcore",
+				"/proc/keys",
+				"/proc/latency_stats",
+				"/proc/timer_list",
+				"/proc/timer_stats",
+				"/proc/sched_debug",
+				"/proc/scsi",
+				"/sys/firmware",
+			} {
+				specgen.AddLinuxMaskedPaths(mp)
+			}
+
+			for _, rp := range []string{
+				"/proc/asound",
+				"/proc/bus",
+				"/proc/fs",
+				"/proc/irq",
+				"/proc/sys",
+				"/proc/sysrq-trigger",
+			} {
+				specgen.AddLinuxReadonlyPaths(rp)
+			}
+		}
+	}
+	// Join the namespace paths for the pod sandbox container.
+	podInfraState := s.Runtime().ContainerStatus(sb.InfraContainer())
+
+	logrus.Debugf("pod container state %+v", podInfraState)
+
+	ipcNsPath := fmt.Sprintf("/proc/%d/ns/ipc", podInfraState.Pid)
+	if err := specgen.AddOrReplaceLinuxNamespace(string(rspec.IPCNamespace), ipcNsPath); err != nil {
+		return nil, err
+	}
+
+	utsNsPath := fmt.Sprintf("/proc/%d/ns/uts", podInfraState.Pid)
+	if err := specgen.AddOrReplaceLinuxNamespace(string(rspec.UTSNamespace), utsNsPath); err != nil {
+		return nil, err
+	}
+
+	if containerConfig.GetLinux().GetSecurityContext().GetNamespaceOptions().GetPid() == pb.NamespaceMode_NODE {
+		// kubernetes PodSpec specify to use Host PID namespace
+		specgen.RemoveLinuxNamespace(string(rspec.PIDNamespace))
+	} else if containerConfig.GetLinux().GetSecurityContext().GetNamespaceOptions().GetPid() == pb.NamespaceMode_POD {
+		// share Pod PID namespace
+		pidNsPath := fmt.Sprintf("/proc/%d/ns/pid", podInfraState.Pid)
+		if err := specgen.AddOrReplaceLinuxNamespace(string(rspec.PIDNamespace), pidNsPath); err != nil {
+			return nil, err
+		}
+	}
+
+	// If the sandbox is configured to run in the host network, do not create a new network namespace
+	if sb.HostNetwork() {
+		if err := specgen.RemoveLinuxNamespace(string(rspec.NetworkNamespace)); err != nil {
+			return nil, err
+		}
+		specgen.RemoveMount("/sys")
+		specgen.RemoveMount("/sys/cgroup")
+		sysMnt := rspec.Mount{
+			Destination: "/sys",
+			Type:        "bind",
+			Source:      "/sys",
+			Options:     []string{"nosuid", "noexec", "nodev", "ro", "rbind"},
+		}
+		specgen.AddMount(sysMnt)
+	} else {
+		netNsPath := sb.NetNsPath()
+		if netNsPath == "" {
+			// The sandbox does not have a permanent namespace,
+			// it's on the host one.
+			netNsPath = fmt.Sprintf("/proc/%d/ns/net", podInfraState.Pid)
+		}
+
+		if err := specgen.AddOrReplaceLinuxNamespace(string(rspec.NetworkNamespace), netNsPath); err != nil {
+			return nil, err
+		}
+	}
+
+	imageSpec := containerConfig.GetImage()
+	if imageSpec == nil {
+		return nil, fmt.Errorf("CreateContainerRequest.ContainerConfig.Image is nil")
+	}
+
+	image := imageSpec.Image
+	if image == "" {
+		return nil, fmt.Errorf("CreateContainerRequest.ContainerConfig.Image.Image is empty")
+	}
+	images, err := s.StorageImageServer().ResolveNames(image)
+	if err != nil {
+		if err == storage.ErrCannotParseImageID {
+			images = append(images, image)
+		} else {
+			return nil, err
+		}
+	}
+
+	// Get imageName and imageRef that are later requested in container status
+	var (
+		imgResult    *storage.ImageResult
+		imgResultErr error
+	)
+	for _, img := range images {
+		imgResult, imgResultErr = s.StorageImageServer().ImageStatus(s.ImageContext(), img)
+		if imgResultErr == nil {
+			break
+		}
+	}
+	if imgResultErr != nil {
+		return nil, imgResultErr
+	}
+
+	imageName := imgResult.Name
+	imageRef := imgResult.ID
+	if len(imgResult.RepoDigests) > 0 {
+		imageRef = imgResult.RepoDigests[0]
+	}
+
+	specgen.AddAnnotation(annotations.Image, image)
+	specgen.AddAnnotation(annotations.ImageName, imageName)
+	specgen.AddAnnotation(annotations.ImageRef, imageRef)
+	specgen.AddAnnotation(annotations.IP, sb.IP())
+
+	// Remove the default /dev/shm mount to ensure we overwrite it
+	specgen.RemoveMount("/dev/shm")
+
+	mnt = rspec.Mount{
+		Type:        "bind",
+		Source:      sb.ShmPath(),
+		Destination: "/dev/shm",
+		Options:     []string{"rw", "bind"},
+	}
+	// bind mount the pod shm
+	specgen.AddMount(mnt)
+
+	options := []string{"rw"}
+	if readOnlyRootfs {
+		options = []string{"ro"}
+	}
+	if sb.ResolvPath() != "" {
+		if err := securityLabel(sb.ResolvPath(), mountLabel, false); err != nil {
+			return nil, err
+		}
+
+		mnt = rspec.Mount{
+			Type:        "bind",
+			Source:      sb.ResolvPath(),
+			Destination: "/etc/resolv.conf",
+			Options:     []string{"bind", "nodev", "nosuid", "noexec"},
+		}
+		// bind mount the pod resolver file
+		specgen.AddMount(mnt)
+	}
+
+	if sb.HostnamePath() != "" {
+		if err := securityLabel(sb.HostnamePath(), mountLabel, false); err != nil {
+			return nil, err
+		}
+
+		mnt = rspec.Mount{
+			Type:        "bind",
+			Source:      sb.HostnamePath(),
+			Destination: "/etc/hostname",
+			Options:     append(options, "bind"),
+		}
+		specgen.AddMount(mnt)
+	}
+
+	if !isInCRIMounts("/etc/hosts", containerConfig.GetMounts()) && hostNetwork(containerConfig) {
+		// Only bind mount for host netns and when CRI does not give us any hosts file
+		mnt = rspec.Mount{
+			Type:        "bind",
+			Source:      "/etc/hosts",
+			Destination: "/etc/hosts",
+			Options:     append(options, "bind"),
+		}
+		specgen.AddMount(mnt)
+	}
+
+	// Set hostname and add env for hostname
+	specgen.SetHostname(sb.Hostname())
+	specgen.AddProcessEnv("HOSTNAME", sb.Hostname())
+
+	specgen.AddAnnotation(annotations.Name, containerName)
+	specgen.AddAnnotation(annotations.ContainerID, containerID)
+	specgen.AddAnnotation(annotations.SandboxID, sb.ID())
+	specgen.AddAnnotation(annotations.SandboxName, sb.InfraContainer().Name())
+	specgen.AddAnnotation(annotations.ContainerType, annotations.ContainerTypeContainer)
+	specgen.AddAnnotation(annotations.LogPath, logPath)
+	specgen.AddAnnotation(annotations.TTY, fmt.Sprintf("%v", containerConfig.Tty))
+	specgen.AddAnnotation(annotations.Stdin, fmt.Sprintf("%v", containerConfig.Stdin))
+	specgen.AddAnnotation(annotations.StdinOnce, fmt.Sprintf("%v", containerConfig.StdinOnce))
+	specgen.AddAnnotation(annotations.ResolvPath, sb.InfraContainer().CrioAnnotations()[annotations.ResolvPath])
+
+	created := time.Now()
+	specgen.AddAnnotation(annotations.Created, created.Format(time.RFC3339Nano))
+
+	metadataJSON, err := json.Marshal(metadata)
+	if err != nil {
+		return nil, err
+	}
+	specgen.AddAnnotation(annotations.Metadata, string(metadataJSON))
+
+	labelsJSON, err := json.Marshal(labels)
+	if err != nil {
+		return nil, err
+	}
+	specgen.AddAnnotation(annotations.Labels, string(labelsJSON))
+
+	kubeAnnotationsJSON, err := json.Marshal(kubeAnnotations)
+	if err != nil {
+		return nil, err
+	}
+	specgen.AddAnnotation(annotations.Annotations, string(kubeAnnotationsJSON))
+
+	spp := containerConfig.GetLinux().GetSecurityContext().GetSeccompProfilePath()
+	if !privileged {
+		if err = s.setupSeccomp(&specgen, spp); err != nil {
+			return nil, err
+		}
+	}
+	specgen.AddAnnotation(annotations.SeccompProfilePath, spp)
+
+	containerIDMappings := s.defaultIDMappings
+
+	metaname := metadata.Name
+	attempt := metadata.Attempt
+	containerInfo, err := s.StorageRuntimeServer().CreateContainer(s.ImageContext(),
+		sb.Name(), sb.ID(),
+		image, imgResult.ID,
+		containerName, containerID,
+		metaname,
+		attempt,
+		mountLabel,
+		containerIDMappings,
+		nil)
+	if err != nil {
+		return nil, err
+	}
+	defer func() {
+		if err != nil {
+			err2 := s.StorageRuntimeServer().DeleteContainer(containerInfo.ID)
+			if err2 != nil {
+				logrus.Warnf("Failed to cleanup container directory: %v", err2)
+			}
+		}
+	}()
+
+	mountPoint, err := s.StorageRuntimeServer().StartContainer(containerID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to mount container %s(%s): %v", containerName, containerID, err)
+	}
+	specgen.AddAnnotation(annotations.MountPoint, mountPoint)
+
+	containerImageConfig := containerInfo.Config
+	if containerImageConfig == nil {
+		err = fmt.Errorf("empty image config for %s", image)
+		return nil, err
+	}
+
+	if containerImageConfig.Config.StopSignal != "" {
+		// this key is defined in image-spec conversion document at https://github.com/opencontainers/image-spec/pull/492/files#diff-8aafbe2c3690162540381b8cdb157112R57
+		specgen.AddAnnotation("org.opencontainers.image.stopSignal", containerImageConfig.Config.StopSignal)
+	}
+
+	// Add image volumes
+	volumeMounts, err := addImageVolumes(mountPoint, s, &containerInfo, &specgen, mountLabel)
+	if err != nil {
+		return nil, err
+	}
+
+	processArgs, err := buildOCIProcessArgs(containerConfig, containerImageConfig)
+	if err != nil {
+		return nil, err
+	}
+	specgen.SetProcessArgs(processArgs)
+
+	envs := mergeEnvs(containerImageConfig, containerConfig.GetEnvs())
+	for _, e := range envs {
+		parts := strings.SplitN(e, "=", 2)
+		specgen.AddProcessEnv(parts[0], parts[1])
+	}
+
+	// Set working directory
+	// Pick it up from image config first and override if specified in CRI
+	containerCwd := "/"
+	if containerImageConfig != nil {
+		imageCwd := containerImageConfig.Config.WorkingDir
+		if imageCwd != "" {
+			containerCwd = imageCwd
+		}
+	}
+	runtimeCwd := containerConfig.WorkingDir
+	if runtimeCwd != "" {
+		containerCwd = runtimeCwd
+	}
+	specgen.SetProcessCwd(containerCwd)
+	if err := setupWorkingDirectory(mountPoint, mountLabel, containerCwd); err != nil {
+		if err1 := s.StorageRuntimeServer().StopContainer(containerID); err1 != nil {
+			return nil, fmt.Errorf("can't umount container after cwd error %v: %v", err, err1)
+		}
+		return nil, err
+	}
+
+	var secretMounts []rspec.Mount
+	if len(s.config.DefaultMounts) > 0 {
+		// This option has been deprecated, once it is removed in the later versions, delete the server/secrets.go file as well
+		logrus.Warnf("--default-mounts has been deprecated and will be removed in future versions. Add mounts to either %q or %q", secrets.DefaultMountsFile, secrets.OverrideMountsFile)
+		var err error
+		secretMounts, err = addSecretsBindMounts(mountLabel, containerInfo.RunDir, s.config.DefaultMounts, specgen)
+		if err != nil {
+			return nil, fmt.Errorf("failed to mount secrets: %v", err)
+		}
+	}
+	// Add secrets from the default and override mounts.conf files
+	secretMounts = append(secretMounts, secrets.SecretMounts(mountLabel, containerInfo.RunDir, s.config.DefaultMountsFile)...)
+
+	mounts := []rspec.Mount{}
+	mounts = append(mounts, ociMounts...)
+	mounts = append(mounts, volumeMounts...)
+	mounts = append(mounts, secretMounts...)
+
+	sort.Sort(orderedMounts(mounts))
+
+	for _, m := range mounts {
+		mnt = rspec.Mount{
+			Type:        "bind",
+			Source:      m.Source,
+			Destination: m.Destination,
+			Options:     append(m.Options, "bind"),
+		}
+		specgen.AddMount(mnt)
+	}
+
+	if err := s.setupOCIHooks(&specgen, sb, containerConfig, processArgs[0]); err != nil {
+		return nil, err
+	}
+
+	// Setup user and groups
+	if linux != nil {
+		if err = setupContainerUser(&specgen, mountPoint, linux.GetSecurityContext(), containerImageConfig); err != nil {
+			return nil, err
+		}
+	}
+
+	// Set up pids limit if pids cgroup is mounted
+	if findCgroupMountpoint("pids") == nil {
+		specgen.SetLinuxResourcesPidsLimit(s.config.PidsLimit)
+	}
+
+	// by default, the root path is an empty string. set it now.
+	specgen.SetRootPath(mountPoint)
+
+	crioAnnotations := specgen.Spec().Annotations
+
+	container, err := oci.NewContainer(containerID, containerName, containerInfo.RunDir, logPath, sb.NetNs().Path(), labels, crioAnnotations, kubeAnnotations, image, imageName, imageRef, metadata, sb.ID(), containerConfig.Tty, containerConfig.Stdin, containerConfig.StdinOnce, sb.Privileged(), sb.Trusted(), containerInfo.RunDir, created, containerImageConfig.Config.StopSignal)
+	if err != nil {
+		return nil, err
+	}
+
+	container.SetIDMappings(containerIDMappings)
+	if s.defaultIDMappings != nil && !s.defaultIDMappings.Empty() {
+		userNsPath := sb.UserNsPath()
+		rootPair := s.defaultIDMappings.RootPair()
+
+		if err := specgen.AddOrReplaceLinuxNamespace(string(rspec.UserNamespace), userNsPath); err != nil {
+			return nil, err
+		}
+		for _, uidmap := range s.defaultIDMappings.UIDs() {
+			specgen.AddLinuxUIDMapping(uint32(uidmap.HostID), uint32(uidmap.ContainerID), uint32(uidmap.Size))
+		}
+		for _, gidmap := range s.defaultIDMappings.GIDs() {
+			specgen.AddLinuxGIDMapping(uint32(gidmap.HostID), uint32(gidmap.ContainerID), uint32(gidmap.Size))
+		}
+		err = s.configureIntermediateNamespace(&specgen, container, sb.InfraContainer())
+		if err != nil {
+			return nil, err
+		}
+
+		if sb.ResolvPath() != "" {
+			err = os.Chown(sb.ResolvPath(), rootPair.UID, rootPair.GID)
+			if err != nil {
+				return nil, err
+			}
+		}
+
+		defer func() {
+			if err != nil {
+				os.RemoveAll(container.IntermediateMountPoint())
+			}
+		}()
+	}
+
+	saveOptions := generate.ExportOptions{}
+	if err = specgen.SaveToFile(filepath.Join(containerInfo.RunDir, "config.json"), saveOptions); err != nil {
+		return nil, err
+	}
+
+	container.SetSpec(specgen.Spec())
+	container.SetMountPoint(mountPoint)
+	container.SetSeccompProfilePath(spp)
+
+	for _, cv := range containerVolumes {
+		container.AddVolume(cv)
+	}
+
+	return container, nil
+}
+
+func setupWorkingDirectory(rootfs, mountLabel, containerCwd string) error {
+	fp, err := symlink.FollowSymlinkInScope(filepath.Join(rootfs, containerCwd), rootfs)
+	if err != nil {
+		return err
+	}
+	if err := os.MkdirAll(fp, 0755); err != nil {
+		return err
+	}
+	if mountLabel != "" {
+		if err1 := securityLabel(fp, mountLabel, false); err1 != nil {
+			return err1
+		}
+	}
+	return nil
+}
+
+func setOCIBindMountsPrivileged(g *generate.Generator) {
+	spec := g.Spec()
+	// clear readonly for /sys and cgroup
+	for i, m := range spec.Mounts {
+		if spec.Mounts[i].Destination == "/sys" {
+			clearReadOnly(&spec.Mounts[i])
+		}
+		if m.Type == "cgroup" {
+			clearReadOnly(&spec.Mounts[i])
+		}
+	}
+	spec.Linux.ReadonlyPaths = nil
+	spec.Linux.MaskedPaths = nil
+}
+
+func clearReadOnly(m *rspec.Mount) {
+	var opt []string
+	for _, o := range m.Options {
+		if o != "ro" {
+			opt = append(opt, o)
+		}
+	}
+	m.Options = append(opt, "rw")
+}
+
+func addOCIBindMounts(mountLabel string, containerConfig *pb.ContainerConfig, specgen *generate.Generator, bindMountPrefix string) ([]oci.ContainerVolume, []rspec.Mount, error) {
+	volumes := []oci.ContainerVolume{}
+	ociMounts := []rspec.Mount{}
+	mounts := containerConfig.GetMounts()
+	for _, mount := range mounts {
+		dest := mount.ContainerPath
+		if dest == "" {
+			return nil, nil, fmt.Errorf("Mount.ContainerPath is empty")
+		}
+
+		if mount.HostPath == "" {
+			return nil, nil, fmt.Errorf("Mount.HostPath is empty")
+		}
+		src := filepath.Join(bindMountPrefix, mount.HostPath)
+
+		resolvedSrc, err := resolveSymbolicLink(src, bindMountPrefix)
+		if err == nil {
+			src = resolvedSrc
+		} else {
+			if !os.IsNotExist(err) {
+				return nil, nil, fmt.Errorf("failed to resolve symlink %q: %v", src, err)
+			} else if err = os.MkdirAll(src, 0644); err != nil {
+				return nil, nil, fmt.Errorf("Failed to mkdir %s: %s", src, err)
+			}
+		}
+
+		options := []string{"rw"}
+		if mount.Readonly {
+			options = []string{"ro"}
+		}
+		options = append(options, "rbind")
+
+		// mount propagation
+		mountInfos, err := dockermounts.GetMounts(nil)
+		if err != nil {
+			return nil, nil, err
+		}
+		switch mount.GetPropagation() {
+		case pb.MountPropagation_PROPAGATION_PRIVATE:
+			options = append(options, "rprivate")
+			// Since default root propagation in runc is rprivate ignore
+			// setting the root propagation
+		case pb.MountPropagation_PROPAGATION_BIDIRECTIONAL:
+			if err := ensureShared(src, mountInfos); err != nil {
+				return nil, nil, err
+			}
+			options = append(options, "rshared")
+			specgen.SetLinuxRootPropagation("rshared")
+		case pb.MountPropagation_PROPAGATION_HOST_TO_CONTAINER:
+			if err := ensureSharedOrSlave(src, mountInfos); err != nil {
+				return nil, nil, err
+			}
+			options = append(options, "rslave")
+			if specgen.Spec().Linux.RootfsPropagation != "rshared" &&
+				specgen.Spec().Linux.RootfsPropagation != "rslave" {
+				specgen.SetLinuxRootPropagation("rslave")
+			}
+		default:
+			logrus.Warnf("Unknown propagation mode for hostPath %q", mount.HostPath)
+			options = append(options, "rprivate")
+		}
+
+		if mount.SelinuxRelabel {
+			if err := securityLabel(src, mountLabel, false); err != nil {
+				return nil, nil, err
+			}
+		}
+
+		volumes = append(volumes, oci.ContainerVolume{
+			ContainerPath: dest,
+			HostPath:      src,
+			Readonly:      mount.Readonly,
+		})
+
+		ociMounts = append(ociMounts, rspec.Mount{
+			Source:      src,
+			Destination: dest,
+			Options:     options,
+		})
+	}
+
+	return volumes, ociMounts, nil
 }
